requirements.txt
flask
python-dotenv

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-pinecone
langchain-groq

pinecone-client
huggingface-hub
requests

pypdf
gunicorn

